+++ src/Type/Definition/Description.php
@@ -7,6 +7,5 @@ class Description
 {
     public function __construct(
         public string $description,
-    ) {
-    }
+    ) {}
 }