+++ CHANGELOG.md
@@ -9,6 +9,12 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v15.2.0
+
+### Changed
+
+- Improve validation messages for type definitions
+
 ## v15.1.0
 
 ### Added